\name{EightSpAbundance}
\docType{data}
\alias{EightSpAbundance}
\title{
  Abundances of 8 species to run examples.
}
\description{
  This dataset is a light-weight example.
}
\usage{
data(Paracou618)
}
\format{
  A named vector.
}

\examples{

data(Paracou618)
EightSpAbundance

}

\keyword{datasets}
