\name{Paracou618.MC}
\docType{data}
\alias{Paracou618.MC}
\title{
  Paracou field station plots 6 and 18, two 1-ha plots inventoried by the Bridge project.
}
\description{
  This dataset is from Paracou field station, French Guiana, managed by \href{http://www.cirad.fr}{Cirad}.
}
\usage{
data(Paracou618)
}
\format{
  An object of class \code{\link{MetaCommunity}} made of two communities and 425 species.
}

\source{
  Permanent data census of Paracou and Marcon et al. (2012).
}
\references{
  Gourlet-Fleury, S., Guehl, J. M. and Laroussinie, O., Eds. (2004). \emph{Ecology & management of a neotropical rainforest. Lessons drawn from Paracou, a long-term experimental research site in French Guiana}. Paris, Elsevier.
  
  Marcon, E., F. Puech, et al. (2012). Characterizing the relative spatial structure of point patterns. \emph{International Journal of Ecology} 2012(Article ID 619281): 11.
}
\examples{

data(Paracou618)
summary(Paracou618.MC)

}

\keyword{datasets}
