\name{EightSpTree}
\docType{data}
\alias{EightSpTree}
\title{
  Functional tree with 8 species.
}
\description{
  This dataset is a light-weight example.
}
\usage{
data(Paracou618)
}
\format{
  An object of class \code{\link{phylog}} containing a functional tree.
}

\examples{
data(Paracou618)
plot(EightSpTree)
}

\keyword{datasets}
